button props interface localized
